Renamed component to better reflect what GitHub API makes available, namely commits per developer
